+++ lib/core/service/firebase_messaging_service.dart
@@ -6,10 +6,10 @@ import 'package:application_base/core/service/service_locator.dart';
 import 'package:firebase_base/core/entity/push_entity.dart';
 import 'package:firebase_base/core/service/local_notifications_service.dart';
 import 'package:firebase_messaging/firebase_messaging.dart';
+import 'package:injectable/injectable.dart';
+import 'package:meta/meta.dart';
 import 'package:rxdart/rxdart.dart';
 
-/// Singleton
-///
 /// Messages have different behaviour depends on application state and OS.
 /// Application state can be:
 ///
@@ -35,7 +35,12 @@ import 'package:rxdart/rxdart.dart';
 /// Details in [Firebase docs](https://firebase.google.com/docs/cloud-messaging/flutter/receive)
 ///
 /// Sheme with common information [here](https://user-images.githubusercontent.com/40064496/197368144-7bfcee7e-644a-4bdc-80f1-b4d38c2eaaff.png)
+@lazySingleton
 final class FirebaseMessagingService with LoggingMixin {
+  ///
+  @visibleForTesting
+  FirebaseMessagingService();
+
   /// Name for logging
   @override
   final String logName = 'Firebase messaging';