+++ lib/core/service/local_notifications_service.dart
@@ -2,14 +2,19 @@ import 'dart:async';
 
 import 'package:awesome_notifications/awesome_notifications.dart';
 import 'package:firebase_base/core/entity/push_entity.dart';
+import 'package:injectable/injectable.dart';
+import 'package:meta/meta.dart';
 
 ///
 typedef HandleMessage = void Function(String? message);
 
-/// Singleton
-///
 /// Used only for Android
+@lazySingleton
 final class LocalNotificationsService {
+  ///
+  @visibleForTesting
+  LocalNotificationsService();
+
   ///
   static const _payloadField = 'payload';
 
@@ -90,8 +95,6 @@ final class LocalNotificationsService {
   /// to the Flutter engine that the dart address will be called from native
   /// and should be preserved
   @pragma('vm:entry-point')
-  static Future<void> _onActionReceivedMethod(
-    ReceivedAction data,
-  ) async =>
+  static Future<void> _onActionReceivedMethod(ReceivedAction data) async =>
       _handleMessage?.call(data.payload?[_payloadField]);
 }